requirements.txt
sqlalchemy[asyncio]
telethon
python-dotenv
aiosqlite
